refactor : client intercepter와 server interceptor를 통합
