+++ apis/config/publicApi.ts
@@ -1,6 +1,5 @@
-import { onResponse, onResponseError } from './interceptor';
 import { BASE_API_URL } from '@/constants';
-import axios from 'axios';
+import axios, { AxiosResponse } from 'axios';
 
 import type { CustomInstance } from './type';
 
@@ -9,6 +8,9 @@ const publicApi: CustomInstance = axios.create({
   withCredentials: true,
 });
 
-publicApi.interceptors.response.use(onResponse, onResponseError);
+publicApi.interceptors.response.use(
+  (response: AxiosResponse) => response.data,
+  (response: AxiosResponse) => response.data
+);
 
 export default publicApi;
