+++ apis/groups/mutations.ts
@@ -1,5 +1,5 @@
-import { useMutation } from '@tanstack/react-query';
 import { postCreateGroup } from './apis.client';
+import { useMutation } from '@tanstack/react-query';
 
 export const usePostCreateGroup = () => {
   return useMutation(postCreateGroup, {});
