+++ apis/groups/queries.ts
@@ -1,6 +1,6 @@
-import { useQuery } from '@tanstack/react-query';
+import { getGroups } from './apis.client';
 import { Keys } from './keys';
-import { getGroups } from './apis.server';
+import { useQuery } from '@tanstack/react-query';
 
 export const useGetGroups = (pageNum: number) => {
   return useQuery(Keys.getGroups(), () => getGroups(pageNum));
