+++ app/page.tsx
@@ -1,17 +1,18 @@
 'use client';
-import { LoginResponse, postLogin, useLoginMutation } from '@/apis/auth';
+import { LoginResponse, useLoginMutation } from '@/apis/auth';
 import { useUser } from '@/hooks/useUser';
+import { getCookie } from '@/utils/cookie';
 
 export default function Home() {
   const { userLogin } = useUser();
   const { mutate: mutateLogin } = useLoginMutation();
 
   const handleGetToken = () => {
+    console.log(getCookie('gloddy_at'));
     mutateLogin(
       { phoneNumber: '010-5728-9357' },
       {
         onSuccess: (response: LoginResponse) => {
-          console.log(response);
           if (response.existUser) {
             const {
               token: { accessToken, refreshToken },
