+++ constants/index.ts
@@ -5,3 +5,4 @@ export const BASE_WEB_URL = process.env.NEXT_PUBLIC_BASE_WEB_URL ?? 'http://loca
 export const INSTAGRAM_URL = 'https://www.instagram.com/gloddykorea';
 
 export const DAY_OF_WEEK = ['SUN', 'MON', 'TUE', 'WED', 'THU', 'FRI', 'SAT'];
+export const isServer = typeof window === undefined;
