+++ utils/auth/tokenValidator.ts
@@ -1,8 +1,8 @@
 import { generateCookiesKeyValues } from '../auth';
 import { postReissue } from '@/apis/auth';
 import { ApiError } from '@/apis/config/customError';
+import { isServer } from '@/constants';
 import { CookieKeyType } from '@/types';
-import { cookies } from 'next/headers';
 
 export const getAccessTokenServer = async (
   authTokens: CookieKeyType
@@ -18,15 +18,25 @@ export const getAccessTokenServer = async (
       ) {
         return null;
       }
-      const cookieStore = cookies();
-      for (const [cookieKey, cookieValue] of generateCookiesKeyValues({
-        accessToken: response.token.accessToken,
-        refreshToken: response.token.refreshToken,
-        userId,
-      })) {
-        cookieStore.set(cookieKey as string, cookieValue as string);
+      if (isServer) {
+        const { cookies } = await import('next/headers');
+        const cookieStore = cookies();
+        for (const [cookieKey, cookieValue] of generateCookiesKeyValues({
+          accessToken: response.token.accessToken,
+          refreshToken: response.token.refreshToken,
+          userId,
+        })) {
+          cookieStore.set(cookieKey as string, cookieValue as string);
+        }
+      } else {
+        for (const [cookieKey, cookieValue] of generateCookiesKeyValues({
+          accessToken: response.token.accessToken,
+          refreshToken: response.token.refreshToken,
+          userId,
+        })) {
+          document.cookie = `${cookieKey}=${cookieValue}; path=/;`;
+        }
       }
-
       return response.token.accessToken;
     } else {
       return null;
